Bug 725: [G50_PVT2] Fail to get sensor status after AC power on.
At boot time kernel will bind devices described in DTS to appropriate
driver, and some drivers will refuse to support a device if that device
doesn't work at that moment.

In the case of G50, sensors on UBB will not work until C2010 powers on,
therefore BMC kernel drivers do not recognize them.

This can be fixed by asking BMC kernel to bind a device to driver when
C2010 powers on. And this binding is done in the same order as defined
in DTS to preserve virtual I2C bus's channel number.

Change-Id: I6c7276720a0882ce7973db3e493d6e2c74e2099a
Signed-off-by: Wei-Chung Wen <wei-chung.wen@foxconn.com>